+++ payzen/classes/admin/PayzenHelperForm.php
@@ -33,6 +33,7 @@ class PayzenHelperForm
         foreach (PayzenApi::getSupportedLanguages() as $code => $label) {
             $languages[$code] = $payzen->l($label, 'payzenhelperform');
         }
+
         asort($languages);
 
         $category_options = array(
@@ -79,7 +80,7 @@ class PayzenHelperForm
         }
 
         $enabledCountries = Country::getCountries((int) $context->language->id, true);
-        $all_countries =    Country::getCountries((int) $context->language->id, false);
+        $all_countries = Country::getCountries((int) $context->language->id, false);
         $countryList = array();
         foreach ($enabledCountries as $value) {
             $countryList['ps_countries'][$value['iso_code']] = $value['name'];
@@ -346,12 +347,6 @@ class PayzenHelperForm
             'PAYZEN_PAYPAL_AMOUNTS' => self::getArrayConfig('PAYZEN_PAYPAL_AMOUNTS'),
             'PAYZEN_PAYPAL_DELAY' => Configuration::get('PAYZEN_PAYPAL_DELAY'),
             'PAYZEN_PAYPAL_VALIDATION' => Configuration::get('PAYZEN_PAYPAL_VALIDATION'),
-
-            'PAYZEN_CHOOZEO_TITLE' => self::getLangConfig('PAYZEN_CHOOZEO_TITLE'),
-            'PAYZEN_CHOOZEO_ENABLED' => Configuration::get('PAYZEN_CHOOZEO_ENABLED'),
-            'PAYZEN_CHOOZEO_AMOUNTS' => self::getArrayConfig('PAYZEN_CHOOZEO_AMOUNTS'),
-            'PAYZEN_CHOOZEO_DELAY' => Configuration::get('PAYZEN_CHOOZEO_DELAY'),
-            'PAYZEN_CHOOZEO_OPTIONS' => self::getArrayConfig('PAYZEN_CHOOZEO_OPTIONS'),
 
             'PAYZEN_OTHER_GROUPED_VIEW' => Configuration::get('PAYZEN_OTHER_GROUPED_VIEW'),
             'PAYZEN_OTHER_ENABLED' => Configuration::get('PAYZEN_OTHER_ENABLED'),